refactor: move definition discovery into Core\Services
DefinitionDiscovery is an internal service of the discovery pipeline, so it
joins ComponentReferenceSigner under Core\Services, dissolving the single-class
Discovery namespace.
